requirements.txt
fastapi==0.109.0
supabase>=2.0.0
python-dotenv==1.0.0
uvicorn[standard]==0.24.0
pydantic>=2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
cryptography==41.0.7